feat: 게시글 추천을 한 기록이 없을 때 예외 처리 추가

+++ src/main/java/balancetalk/module/post/application/PostService.java
@@ -152,6 +152,7 @@ public class PostService {
             throw new BalanceTalkException(ALREADY_LIKE_POST);
         }
 
+
         PostLike postLike = PostLike.builder()
                 .post(post)
                 .member(member)
@@ -164,9 +165,15 @@ public class PostService {
     public void cancelLikePost(Long postId) {
         Post post = getCurrentPost(postId);
         Member member = getCurrentMember(memberRepository);
+        if (notExistsPostLikeBy(member, post)) {
+            throw new BalanceTalkException(NOT_FOUND_POST_LIKE);
+        }
         postLikeRepository.deleteByMemberAndPost(member, post);
     }
 
+    private boolean notExistsPostLikeBy(Member member, Post post) {
+        return !postLikeRepository.existsByMemberAndPost(member, post);
+    }
     private Post getCurrentPost(Long postId) {
         return postRepository.findById(postId)
                 .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_POST));
